Crafting simulation performance patches (#525)
* Remove stream API usage from CraftableItemResolver

Streams are not that great on hotpaths...

* Only copy array when any elements were removed while iterating in IItemList.toArray

Previously, the ArrayLIst used allocated an array internally and then always copied it again when returning from this method. The second copy would unnecessary if no elements were filtered out by the iterator

* Add some capacity adjustments for local ArrayLists in CraftableItemResolver

* Replace providers Map with a simpler linear access time Map implementation

The Map will currently only ever have two keys and a maximum of six values for a key. The way it is currently being iterated over imposes an insane overhead. This change tries to replace that overhead with a simple array loop.

* Replace byteAmountAdjusters Map with a simpler linear access time Map implementation

The Map will currently only ever have one key with one value. The way it is currently being iterated over imposes an insane overhead. This change tries to replace that overhead with a simple array loop.

* Use ICraftingPatternDetails#getCondensed{Inputs,Outputs} to remove unnecessary work in CraftFromPatternTask constructor

* Use fastutil Set instead of Map in HashBasedItemList and FluidList

* Add a Set to speedup non fuzzy accesses to ItemList

This greatly speeds up adding to existing items and finding precise items, which are the most used operations in big AE systems. While fully removing items and adding completely new items is slower, these operations are lesser used, therefore making the trade-off worth it.
Also increases memory usage of ItemLists a bit.

* Adjust name of ArrayBasedMultiMap and add tests

* Move test to correct package

* Replace custom map with fastutil ArrayMap